:bug: fix fetching selectedDataset and refactor
