--- src/components/result/CardUtils.tsx
@@ -1,9 +1,0 @@
-const trimContent = (c: string | undefined, size: number = 90): string => {
-  if (c) {
-    return `${c.slice(0, 90)}${c.length > size ? "..." : ""}`;
-  } else {
-    return "";
-  }
-};
-
-export { trimContent };
